Add SecondAbility and SecondMissile, update character and projectile files
Implemented a new ability, SecondAbility, which spawns a new type of missile, SecondMissile. Several changes were made to the character and ability files to allow for the activation of this new ability. The unrelated Tick function from AmathProjectile was removed to streamline the code. Finally, an AMathProjectileAbility was also added, which activates when the corresponding input is received.