Implement character death mechanics and math question widget
Several changes are introduced in this commit. Character death and responses to health depletion are implemented, both for the main character and enemy characters. Additionally, a new 'MathQuestion' Widget Component is added to the main character to support generation of random math quizzes. Other minor tweaks are applied across the code to support and improve these primary changes.